third_party/WebKit/LayoutTests/fast/viewport/viewport-83-expected.txt

CONSOLE ERROR: line 3: The value "height" for key "width" is invalid, and has been ignored.
CONSOLE ERROR: line 3: The key "830" is not recognized and ignored.
ALERT: viewport size 64x70.4 scale 5 with limits [5, 5] and userScalable true
